Refactor: AuthHolder와 AuthUtils를 util로 분류, 이에 따른 코드 수정사항
