+++ src/main/java/com/github/supercodingfinalprojectbackend/service/Oauth2Service.java
@@ -7,7 +7,8 @@ import com.github.supercodingfinalprojectbackend.entity.type.SocialPlatformType;
 import com.github.supercodingfinalprojectbackend.entity.type.UserRole;
 import com.github.supercodingfinalprojectbackend.exception.errorcode.ApiErrorCode;
 import com.github.supercodingfinalprojectbackend.repository.*;
-import com.github.supercodingfinalprojectbackend.util.AuthUtils;
+import com.github.supercodingfinalprojectbackend.util.auth.AuthHolder;
+import com.github.supercodingfinalprojectbackend.util.auth.AuthUtils;
 import com.github.supercodingfinalprojectbackend.util.jwt.JwtUtils;
 import com.github.supercodingfinalprojectbackend.util.jwt.TokenHolder;
 import lombok.RequiredArgsConstructor;
