+++ src/main/java/com/github/supercodingfinalprojectbackend/util/auth/AuthHolder.java
@@ -1,4 +1,4 @@
-package com.github.supercodingfinalprojectbackend.dto;
+package com.github.supercodingfinalprojectbackend.util.auth;
 
 import java.util.concurrent.ConcurrentHashMap;
